examples/shopping-cart/cart/serializers.py
from rest_framework import serializers
from cart.models import Cart, Item


class ItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Item
        fields = ('name', 'quantity',)


class CartSerializer(serializers.ModelSerializer):
    items = ItemSerializer(many=True)

    class Meta:
        model = Cart
        fields = ('items',)
