examples/shopping-cart/requirements.txt
Django
djangorestframework
edn-format==0.5.12
git+https://github.com/jstaffans/django-cljs-loader.git@7cca424e94377ad046847a257a787c2f3cab1665#egg=django-cljs-loader
